added uploader, changed some files
added carrierwave uploader "photo" for the images table, created a migration to add a photo field (of type string) to the table images, modified the images controller to add a photo field, and mounted the uploader to the image model file. also added carrierwave to the gemfile. photo_uploader is already set to whitelist image files as uploads only